Added GMTracker support class to help with testing scenes individually:
When loading scene, this class will check if there is an instance of GameManager object exists and create one if doesn't.
It then calls SceneLoaded() method to simulate natural scene transition.

This is for testing purposes. The final build will have one instance of GameManager that will persist through every scene.